[package, tsconfig] Flatten build
Use one toplevel package.json, since TypeScript now supports "composite"
packages. The main reason for this change is so that an `F5` in vscode
displays errors with non-broken links, and `F8` can be used to cycle
through errors quite easily. Without this patch, vscode used to display
paths to files relative to the 'server' and 'client' directories, and
they were broken.

We have dropped html_views in this patch, which the next patch will
address.