Rank possible completions
Ranks based upon atomic type overlap.
Gets the atomic types of goal and lemmas.
Then checks for the largest overlap between the lemma and goal.
This is first done by seeing which most types in common with the goal, if they have an equal amount, then the smallest of the two is used.
By doing it this way it should show lemmas that has the most possible types that are relevant to the current goal, while not showing gigantic lemmas.

A draw back with this method is that it does not show very general lemmas, as they will have a very small overlap.