new logging (not based on Coq's feedback)
initialization (OCaml + vscoqtop) logging is saved to a file
  /tmp/vscoq_init_log.*.txt
since one cannot log anything before vscode is initialized.

if initialization works, then the vscoqtop initialization log
is also reported to the output panel in vscode.

ix logging once and forall (famous last words)

All log info is reported in the output panel.

A prefix of the log is also saved in /tmp/vscoq_log*txt, in case
the server crashes to early to write to the panel (eg during
initialization)