Using interpretToPoint to update goal view in continuous mode.
Now the SendProofView event takes a position so to know where to make
the proof from. Additionally, a ~stateful label was added to the
interpret_to_position function to indicate wether or not to update
observe_id. The background label is kept so to keep the execution
priorities in order in continuous mode.