Run `vscoq.path` as a shell command from the project directory
This fixes an inconsistency between using `child_process.exec`, which runs the provided string as a shell command, and `vscode-languageclient`, which by default did not.

Also, `vscoqtop` is now always run from the directory of the project.

All that allows to use VsCoq for Coq projects built with dune, by putting the following to `.vscode/settings.json` (where `theories/Theory.v` is some Coq file):
```json
{
    "vscoq.path": "dune coq top --toplevel vscoqtop theories/Theory.v --"
}
```