fix: block on parse error
We create a Block type in the scheduler (and its corresponding PBlock
in execution manager) which will send an exec_error if one tries to execute
through a parse error. If the user has set block_on_error mode then the behaviour
is now identical with parsing errors and execution errors.